[setup] wait for Vault to elect a new leader after restart
Vault recognizes that the old leader is still a leader after
restarting all Vault servers.  This is probablly because the
leader information is stored in etcd and each server references it.

What's bad is, while no leader is elected, Vault's Go API client
would returns (nil, nil) for Logical().Write() when issuing
certificates.

To workaround the problem, waitVaultLeader() waits a new leader
is elected.