[ignition] replace neco.target with network-online.target
network-online.target is a pre-installed target for systemd.
The purpose of the target is to wait for completion of all
network configurations.

However, the target was not used; instead neco.target was
installed and used for the same purpose.

This commit removes neco.target and simply replace it with
the pre-installed network-online.target.