Purge NOTRACK config in ignition
On our staging environment, DNS resolution requests from outside of the k8s
cluster failed because the NOTRACK rule prevents conntrack from NATing
the response for the requests.
NOTRACK can be purged because Pod to Node connection does not relate to
conntrack and the worker nodes do DNS resolution via TCP.
This PR purges NOTRACK config and fixes the issue.

Signed-off-by: H.Muraoka <hiroshi-muraoka@cybozu.co.jp>