dctest: dcssh: pass identify file path via command line option
We cannot use `dcssh` and `dcscp` unless the current working directory
is `/path/to/neco/dctest` since the ssh config that these tools use
requires the identity file `dctest_key` exists in the current working
directory. This commit changes `dcssh` and `dcscp` to pass the absolute
path of `dctest_key` to an inner ssh command so that these scripts can
be run from everywhere.