Ignore new artifacts temporarily
Currently the upgrade test never succeed because the libssl package
has been moved.
This branch will fix it, but the build test for this branch fails
due to new artifacts.
I minimize this branch by ignoring new artifacts because this branch
should be merged without the success of the upgrade test.

Signed-off-by: morimoto-cybozu <kenji_morimoto@cybozu.co.jp>