Refine BIRD configuration parameters
This commit updates the BIRD configuration to extend the BFD TX/RX intervals
from 300 ms to 400 ms.
The multiplier is set to 5 in default.
This combination means that a link is not declared as down if the link
failure is shorter than 2 sec.
In other words, the detection of the link failure is delayed by 2 sec.

This commit also shorten the minimum delay between a protocol failure
and automatic restart for the BGP protocols from 60 sec (default) to
3 sec.
This will speed up the recovery from a failure.

Signed-off-by: morimoto-cybozu <kenji_morimoto@cybozu.co.jp>